+++ src/com/Controller/Reports/MasterItemsController.java
@@ -147,7 +147,7 @@ public class MasterItemsController implements Initializable {
     
    private void setItems() throws SQLException{
        String totJan = "0.00", totFeb = "0.00", totMar = "0.00", totApr = "0.00", totMay = "0.00", totJun = "0.00", totJul = "0.00", totAug = "0.00", totSep = "0.00", totOct = "0.00", totNov = "0.00", totDec = "0.00", summation = "0.00", Jan = "0.00", Feb = "0.00", Mar = "0.00", Apr = "0.00", May = "0.00", Jun = "0.00", Jul = "0.00", Aug = "0.00", Sep = "0.00", Oct = "0.00", Nov = "0.00", Dec = "0.00", totalAmnt = "0.00";
-       float  totjan = 0, totfeb = 0,totmar = 0,totapr = 0, totmay = 0, totjun = 0, totjul = 0, totaug = 0,totsep = 0,totoct = 0, totnov = 0, totdec = 0, totMonthsum = 0,  jan = 0, feb = 0, apr = 0, mai = 0, jun = 0, jul = 0, aug = 0, sep = 0, oct = 0, mar = 0, nov = 0, dec = 0, total_amount = 0;
+       double  totjan = 0, totfeb = 0,totmar = 0,totapr = 0, totmay = 0, totjun = 0, totjul = 0, totaug = 0,totsep = 0,totoct = 0, totnov = 0, totdec = 0, totMonthsum = 0,  jan = 0, feb = 0, apr = 0, mai = 0, jun = 0, jul = 0, aug = 0, sep = 0, oct = 0, mar = 0, nov = 0, dec = 0, total_amount = 0;
        NumberFormat formatter = new DecimalFormat("#,##0.00");
        PreparedStatement stmnt_itemsCategories;
        ResultSet rs, rs_itemsCategories;
@@ -196,11 +196,11 @@ public class MasterItemsController implements Initializable {
            revenueItem.setStyle("-fx-alignment: CENTER; -fx-text-fill: #5a5959;");
            getReport = new GetItemsReport(category, "", "", "", "", "", "", "", "", "", "", "", "", "");
            revenueItemsTable.getItems().add(getReport);
-           float subjan = 0, subfeb = 0, subapr = 0, submai = 0, subjun = 0, subjul = 0, subaug = 0, subsep = 0, suboct = 0, submar = 0, subnov = 0, subdec = 0, subtotal_amount;
+           double subjan = 0, subfeb = 0, subapr = 0, submai = 0, subjun = 0, subjul = 0, subaug = 0, subsep = 0, suboct = 0, submar = 0, subnov = 0, subdec = 0, subtotal_amount;
            String subJan = "0.00", subFeb = "0.00", subMar = "0.00", subApr = "0.00", subMay = "0.00", subJun = "0.00", subJul = "0.00", subAug = "0.00", subSep = "0.00", subOct = "0.00", subNov = "0.00", subDec = "0.00", subTotalAmnt = "0.00";
            for (String item : categoriesItem.get(category)) {
-               Map<String, Map<String, Float>> itemMonthSum = new HashMap<>();
-               Map<String, Float> monthSum = new HashMap<>();
+               Map<String, Map<String, Double>> itemMonthSum = new HashMap<>();
+               Map<String, Double> monthSum = new HashMap<>();
                monthSum.put(january.getText(), jan); monthSum.put(february.getText(), feb); monthSum.put(march.getText(), mar);
                monthSum.put(april.getText(), apr); monthSum.put(may.getText(), mai); monthSum.put(june.getText(), jun);
                monthSum.put(july.getText(), jul); monthSum.put(august.getText(), aug); monthSum.put(september.getText(), sep);
@@ -210,8 +210,8 @@ public class MasterItemsController implements Initializable {
                while (resultSetState){
                    rs.next();
                    if (item.equals(rs.getString("revenue_item"))){
-                       float amot= itemMonthSum.get(item).get(Months.get(rs.getInt("revenueMonth")).toString());
-                       amot += rs.getFloat("revenueAmount");
+                       double amot= itemMonthSum.get(item).get(Months.get(rs.getInt("revenueMonth")).toString());
+                       amot += rs.getDouble("revenueAmount");
                        itemMonthSum.get(item).put(Months.get(rs.getInt("revenueMonth")).toString(), amot);
                    }
                    if (rs.isLast()){
@@ -250,13 +250,13 @@ public class MasterItemsController implements Initializable {
     } 
    
    
-       public Float setCentersSum(String Item, String Month, String Year) throws SQLException{
-        float totalAmunt;
+       public Double setCentersSum(String Item, String Month, String Year) throws SQLException{
+        double totalAmunt;
        stmnt = con.prepareStatement(" SELECT `revenueAmount`   FROM `revenue_items`,`daily_entries` WHERE `revenue_item_ID` = `revenueItem` AND `revenue_item` = '"+Item+"' AND `revenueMonth` = '"+Month+"' AND `revenueYear` = '"+Year+"'  ");
        ResultSet rs = stmnt.executeQuery();
-       ObservableList<Float> Amount = FXCollections.observableArrayList();//List to Store revenue items which have entries for the specified week
+       ObservableList<Double> Amount = FXCollections.observableArrayList();//List to Store revenue items which have entries for the specified week
        while(rs.next()){//looping through the retrieved revenueItems result set
-           Amount.add(rs.getFloat("revenueAmount"));
+           Amount.add(rs.getDouble("revenueAmount"));
        }
         totalAmunt = 0;
         for(int i = 0; i < Amount.size(); i++){
@@ -349,8 +349,8 @@ public class MasterItemsController implements Initializable {
  *                 }
  *             }
  *         }
- *           Map<String, ArrayList<Float>> monthAmount = new HashMap<>();//HashMap to store revenue Amounts on their respective weeks
- *           Map<String, Map<String, ArrayList<Float>>> forEntry = new HashMap<>();//HashMap to store entries for tableview
+ *           Map<String, ArrayList<Double>> monthAmount = new HashMap<>();//HashMap to store revenue Amounts on their respective weeks
+ *           Map<String, Map<String, ArrayList<Double>>> forEntry = new HashMap<>();//HashMap to store entries for tableview
  *           rowItems.forEach((Center) -> {
  *               forEntry.put(Center, new HashMap<>());
  *       });
@@ -360,10 +360,10 @@ public class MasterItemsController implements Initializable {
  *           try {
  *           for(String month : rowMonths) {
  *               for(String Item : rowItems) {
- *                   float monthSum;
+ *                   double monthSum;
  *                   monthSum = setCentersSum(Item, month, cmbMasterItemsYear.getSelectionModel().getSelectedItem());
- *                   for(Map.Entry<String, ArrayList<Float>> Dates : monthAmount.entrySet()){
- *                           for(Map.Entry<String, Map<String, ArrayList<Float>>>Items : forEntry.entrySet()){
+ *                   for(Map.Entry<String, ArrayList<Double>> Dates : monthAmount.entrySet()){
+ *                           for(Map.Entry<String, Map<String, ArrayList<Double>>>Items : forEntry.entrySet()){
  *                               if (Items.getKey().equals(Item)  && Dates.getKey().equals(month)){
  *                                   if(forEntry.containsKey(Item) && !forEntry.get(Item).containsKey(month)){
  *                                       forEntry.get(Item).put(month, new ArrayList<>());
@@ -383,11 +383,11 @@ public class MasterItemsController implements Initializable {
  *                   }
  *           NumberFormat formatter = new DecimalFormat("#,##0.00");
  *
- *        for(Map.Entry<String, Map<String, ArrayList<Float>>>Items : forEntry.entrySet()){
+ *        for(Map.Entry<String, Map<String, ArrayList<Double>>>Items : forEntry.entrySet()){
  *            String jan1 = "0.00", feb1 = "0.00", apr1 = "0.00", mai1 = "0.00", jun1 = "0.00", jul1 = "0.00", aug1 = "0.00", sep1 = "0.00",
  *                    oct1 = "0.00", mar1 = "0.00", nov1 = "0.00", dec1 = "0.00", totalAmnt = "0.00";
- *            float jan = 0, feb = 0, apr = 0, mai = 0, jun = 0, jul = 0, aug = 0, sep = 0, oct = 0, mar = 0, nov = 0, dec = 0, total_amount;
- *            for(Map.Entry<String, ArrayList<Float>> Dates :forEntry.get(Items.getKey()).entrySet() ){
+ *            double jan = 0, feb = 0, apr = 0, mai = 0, jun = 0, jul = 0, aug = 0, sep = 0, oct = 0, mar = 0, nov = 0, dec = 0, total_amount;
+ *            for(Map.Entry<String, ArrayList<Double>> Dates :forEntry.get(Items.getKey()).entrySet() ){
  *                String reveItem = Items.getKey();
  *                System.out.println(reveItem+ "\n"+Items.getValue().get(Dates.getKey()));
  *                if(Dates.getKey() == null ? january.getText() == null : Dates.getKey().equals(january.getText())){
@@ -462,13 +462,13 @@ public class MasterItemsController implements Initializable {
  *     }
  *
  *
- *        public Float setCentersSum(String Item, String Month, String Year) throws SQLException{
- *         float totalAmunt;
+ *        public Double setCentersSum(String Item, String Month, String Year) throws SQLException{
+ *         double totalAmunt;
  *        stmnt = con.prepareStatement(" SELECT `revenueAmount`   FROM `revenue_items`,`daily_entries` WHERE `revenue_item_ID` = `revenueItem` AND `revenue_item` = '"+Item+"' AND `revenueMonth` = '"+Month+"' AND `revenueYear` = '"+Year+"'  ");
  *        ResultSet rs = stmnt.executeQuery();
- *        ObservableList<Float> Amount = FXCollections.observableArrayList();//List to Store revenue items which have entries for the specified week
+ *        ObservableList<Double> Amount = FXCollections.observableArrayList();//List to Store revenue items which have entries for the specified week
  *        while(rs.next()){//looping through the retrieved revenueItems result set
- *            Amount.add(rs.getFloat("revenueAmount"));
+ *            Amount.add(rs.getDouble("revenueAmount"));
  *        }
  *         totalAmunt = 0;
  *         for(int i = 0; i < Amount.size(); i++){
